+++ src/contentTypeConfig.js
@@ -154,11 +154,12 @@ export const individualLessonsTypes = [
   ...studentArchivesLessonTypes
 ];
 
-export const coursesLessonTypes = ['course', 'pack','spotlight'];
+export const coursesLessonTypes = ['course', 'pack','spotlight', 'guided-course'];
 export const showsLessonTypes = ['diy-drum-experiment','exploring-beats','in-rhythm',  'rhythmic-adventures-of-captain-carson','rhythms-from-another-planet','study-the-greats'];
 export const collectionLessonTypes = [
     ...coursesLessonTypes,
-    ...showsLessonTypes
+    ...showsLessonTypes,
+    'song-tutorial',
 ];
 
 export const lessonTypesMapping = {
@@ -187,7 +188,7 @@ export const progressTypesMapping = {
   'song tutorial': tutorialsLessonTypes,
   'songs': transcriptionsLessonTypes,
   'play-along': playAlongLessonTypes,
-  'guided course': ['challenge'],
+  'guided course': ['guided-course'],
   'pack': ['pack', 'semester-pack'],
   'method': ['learning-path']
 };
@@ -426,8 +427,8 @@ export let contentTypeConfig = {
             }`,
       `"resources": ${resourcesField}`,
       '"thumbnail": thumbnail.asset->url',
-      '"light_logo": light_mode_logo_url.asset->url',
-      '"dark_logo": dark_mode_logo_url.asset->url',
+      '"light_mode_logo": light_mode_logo_url.asset->url',
+      '"dark_mode_logo": dark_mode_logo_url.asset->url',
       `"description": ${descriptionField}`,
     ],
   },
@@ -449,8 +450,8 @@ export let contentTypeConfig = {
       `"resources": ${resourcesField}`,
       '"image": logo_image_url.asset->url',
       '"thumbnail": thumbnail.asset->url',
-      '"light_logo": light_mode_logo_url.asset->url',
-      '"dark_logo": dark_mode_logo_url.asset->url',
+      '"light_mode_logo": light_mode_logo_url.asset->url',
+      '"dark_mode_logo": dark_mode_logo_url.asset->url',
       `"description": ${descriptionField}`,
       'total_xp',
     ],