+++ src/services/sanity.js
@@ -888,11 +888,11 @@ async function getProgressFilter(progress, progressIds) {
       return `&& !(railcontent_id in [${ids.join(',')}])`
     }
     case 'recent': {
-      const ids = await getAllStartedOrCompleted()
+      const ids = progressIds !== undefined ? progressIds : await getAllStartedOrCompleted()
       return `&& (railcontent_id in [${ids.join(',')}])`
     }
     case 'incomplete': {
-      const ids = await getAllStarted()
+      const ids = progressIds !== undefined ? progressIds :await getAllStarted()
       return `&& railcontent_id in [${ids.join(',')}]`
     }
     default:
@@ -1269,7 +1269,6 @@ export async function fetchLessonContent(railContentId) {
           "id":railcontent_id,
           slug, artist->,
           "thumbnail":thumbnail.asset->url,
-          "url": web_url_path,
           soundslice_slug,
           "description": description[0].children[0].text,
           "chapters": chapter[]{
@@ -1299,9 +1298,11 @@ export async function fetchLessonContent(railContentId) {
           "parent_content_data": parent_content_data[]{
             "id": id,
             "title": *[railcontent_id == ^.id][0].title,
-            "web_url_path": *[railcontent_id == ^.id][0].web_url_path,
             "slug":*[railcontent_id == ^.id][0].slug,
             "type": *[railcontent_id == ^.id][0]._type,
+            "logo" : *[railcontent_id == ^.id][0].logo_image_url.asset->url,
+            "dark_mode_logo": *[railcontent_id == ^.id][0].dark_mode_logo_url.asset->url,
+            "light_mode_logo": *[railcontent_id == ^.id][0].light_mode_logo_url.asset->url,
           },
           sort,
           xp,
@@ -1437,7 +1438,7 @@ export async function fetchRelatedLessons(railContentId, brand) {
     pullFutureContent: true,
   }).buildFilter()
   const childrenFilter = await new FilterBuilder(``, { isChildrenFilter: true }).buildFilter()
-  const queryFields = `_id, "id":railcontent_id, published_on, "instructor": instructor[0]->name, title, "thumbnail":thumbnail.asset->url, length_in_seconds, web_url_path, "type": _type, difficulty, difficulty_string, railcontent_id, artist->,"permission_id": permission[]->railcontent_id,_type, "genre": genre[]->name`
+  const queryFields = `_id, "id":railcontent_id, published_on, "instructor": instructor[0]->name, title, "thumbnail":thumbnail.asset->url, length_in_seconds, status, "type": _type, difficulty, difficulty_string, railcontent_id, artist->,"permission_id": permission[]->railcontent_id,_type, "genre": genre[]->name`
   const queryFieldsWithSort = queryFields + ', sort'
   const query = `*[railcontent_id == ${railContentId} && brand == "${brand}" && (!defined(permission) || references(*[_type=='permission']._id))]{
    _type, parent_type, 'parent_id': parent_content_data[0].id, railcontent_id,
@@ -2187,12 +2188,12 @@ async function buildQuery(
 function buildEntityAndTotalQuery(
   filter = '',
   fields = '...',
-  { sortOrder = 'published_on desc', start = 0, end = 10, isSingle = false }
+  { sortOrder = 'published_on desc', start = 0, end = 10, isSingle = false, withoutPagination = false }
 ) {
-  const sortString = sortOrder ? `order(${sortOrder})` : ''
-  const countString = isSingle ? '[0...1]' : `[${start}...${end}]`
+  const sortString = sortOrder ? ` | order(${sortOrder})` : ''
+  const countString = isSingle ? '[0...1]' : (withoutPagination ? ``: `[${start}...${end}]`)
   const query = `{
-      "entity": *[${filter}] | ${sortString}${countString}
+      "entity": *[${filter}]  ${sortString}${countString}
       {
         ${fields}
       },
@@ -2311,16 +2312,31 @@ export async function fetchTabData(
 ) {
   const start = (page - 1) * limit
   const end = start + limit
-
+  let withoutPagination = false
   // Construct the included fields filter, replacing 'difficulty' with 'difficulty_string'
   const includedFieldsFilter =
     includedFields.length > 0 ? filtersToGroq(includedFields, [], pageName) : ''
 
+  let sortOrder = getSortOrder(sort, brand, '')
+
+  switch (progress) {
+    case 'recent':
+      progressIds = await getAllStartedOrCompleted({ brand, onlyIds: true });
+      sortOrder = null;
+      withoutPagination = true;
+      break;
+    case 'incomplete':
+      progressIds = await getAllStarted();
+      sortOrder = null;
+      break;
+    case 'completed':
+      progressIds = await getAllCompleted();
+      sortOrder = null;
+      break;
+  }
+
   // limits the results to supplied progressIds for started & completed filters
   const progressFilter = await getProgressFilter(progress, progressIds)
-
-  // Determine the sort order
-  const sortOrder = getSortOrder(sort, brand, '')
 
   let fields = DEFAULT_FIELDS
   let fieldsString = fields.join(',')
@@ -2348,9 +2364,23 @@ export async function fetchTabData(
     sortOrder: sortOrder,
     start: start,
     end: end,
+    withoutPagination: withoutPagination,
   })
 
-  return fetchSanity(query, true)
+  let results = await fetchSanity(query, true);
+
+  if (['recent', 'incomplete', 'completed'].includes(progress) && results.entity.length > 1) {
+    const orderMap = new Map(progressIds.map((id, index) => [id, index]))
+    results.entity = results.entity
+      .sort((a, b) => {
+        const aIdx = orderMap.get(a.id) ?? Number.MAX_SAFE_INTEGER;
+        const bIdx = orderMap.get(b.id) ?? Number.MAX_SAFE_INTEGER;
+        return aIdx - bIdx || new Date(b.published_on) - new Date(a.published_on);
+      })
+      .slice(start, end);
+  }
+
+  return results;
 }
 
 export async function fetchRecent(